requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
asyncpg==0.31.0
attrs==26.1.0
Automat==25.4.16
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
comm==0.2.3
constantly==23.10.4
cryptography==46.0.6
cssselect==1.4.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
itemadapter==0.13.1
itemloaders==1.4.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-huggingface==1.2.1
langchain-ollama==1.1.0
langchain-postgres==0.0.17
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.29
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
ollama==0.6.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
parsel==1.11.0
parso==0.8.6
pexpect==4.9.0
pgvector==0.3.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.6.0
psutil==7.2.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyDispatcher==2.0.7
Pygments==2.20.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.9.0
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
Scrapy==2.14.2
sentence-transformers==5.3.0
service-identity==24.2.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.1
triton==3.6.0
Twisted==25.5.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
w3lib==2.4.1
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
zope.interface==8.2
